refactor(datasets): extract master-data routes + ES-error helper to routes/*.ts (6d)
Also fixes permissions.middleware to type its req param as base Request (it only
reads context via accessors), removing the RequestHandler overload friction that
recurs across the datasets route files.
