+++ dev/type-errors-baseline.txt
@@ -1,1 +1,1 @@
-888
+834
